Adding doc + some small optimisations
