+++ TestcsFastFloat/Basic/TestDoubleParser.cs
@@ -9,8 +9,6 @@ using Xunit;
 
 namespace TestcsFastFloat.Tests.Basic
 {
-#pragma warning disable xUnit1026
-
   public class TestDoubleParser : BaseTestClass
   {
     [InlineData("a", "A", 1, true)]
@@ -20,9 +18,8 @@ namespace TestcsFastFloat.Tests.Basic
     unsafe public void Test1(string a, string b, int len, bool res)
     {
       fixed (char* pa = a)
-      fixed (char* pb = b)
       {
-        Assert.Equal(res, Utils.strncasecmp(pa, pb, len));
+        Assert.Equal(res, Utils.strncasecmp(pa, b, len));
       }
     }
 
@@ -45,56 +42,7 @@ namespace TestcsFastFloat.Tests.Basic
       }
     }
 
-    [Fact]
-    unsafe public void ParseNumberString_Works_Scnenarios()
-    {
-      Dictionary<string, string> sut = new Dictionary<string, string>();
-
-      sut.Add("leading zeros", "001");
-      sut.Add("leading zeros neg", "-001");
-
-      sut.Add("zero", "0");
-      sut.Add("zero neg", "-0");
-
-      sut.Add("double", "0.00000000000000212312312");
-      sut.Add("double neg", "-0.00000000000000212312312");
-      sut.Add("int", "1");
-      sut.Add("int neg", "-1");
-
-      sut.Add("autreint ", "123124");
-      sut.Add("autreint neg", "-123124");
-
-      sut.Add("notation scientifique", "4.56E+2");
-      sut.Add("notation scientifique neg", "-4.56E-2");
-
-      sut.Add("notation scientifique 2", "4.5644E+2");
-      sut.Add("notation scientifique 2 neg", "-4.5644E-2");
-
-      sut.Add("notation scientifique 3", "4424.5644E+22");
-      sut.Add("notation scientifique 3 neg", "-4424.5644E-22");
-
-      sut.Add("notation scientifique 4", "4424.5644E+223");
-      sut.Add("notation scientifique 4 neg", "-4424.5644E-223");
-      StringBuilder sb = new StringBuilder();
-
-      foreach (KeyValuePair<string, string> kv in sut)
-      {
-        sb.AppendLine($"Scenario : {kv.Key} ");
-        sb.AppendLine($"Valeur   : {kv.Value} ");
-
-        fixed (char* p = kv.Value)
-        {
-          char* pend = p + kv.Value.Length;
-          var res = ParsedNumberString.ParseNumberString(p, pend);
-
-          sb.AppendLine($"Resultat : {res.exponent} {res.mantissa} {res.negative} {res.valid}");
-          sb.AppendLine();
-        }
-      }
-
-      VerifyData(sb.ToString());
-    }
-
+   
     [Fact]
     unsafe public void ParseNumber_Works_Scnenarios()
     {
