+++ csFastFloat/FastFloatParser.cs
@@ -52,18 +52,24 @@ namespace csFastFloat
     }
 
 
+    /// <summary>
+    /// Parse float from string input
+    /// </summary>
+    /// <param name="s"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     public static unsafe float ParseFloat(string s, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
-    {
-      if (s == null)
-        ThrowArgumentNull();
-        static void ThrowArgumentNull() => throw new ArgumentNullException(nameof(s));
+      => ParseFloat(s, out long _, expectedFormat, decimal_separator);
 
-      fixed (char* pStart = s)
-      {
-        return ParseFloat(pStart, pStart + (uint)s.Length, expectedFormat, decimal_separator);
-      }
-    }
-
+    /// <summary>
+    /// Parse float from string input and count number of characters consumed
+    /// </summary>
+    /// <param name="s"></param>
+    /// <param name="characters_consumed"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     public static unsafe float ParseFloat(string s, out long characters_consumed, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
     {
       if (s == null)
@@ -76,13 +82,24 @@ namespace csFastFloat
       }
     }
 
+
+    /// <summary>
+    /// Parse float from ReadonlySpan of char
+    /// </summary>
+    /// <param name="s"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     public static unsafe float ParseFloat(ReadOnlySpan<char> s, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
-    {
-      fixed (char* pStart = s)
-      {
-        return ParseFloat(pStart, pStart + (uint)s.Length, expectedFormat, decimal_separator);
-      }
-    }
+      => ParseFloat(s, out long _, expectedFormat, decimal_separator);
+    /// <summary>
+    /// Parse float from ReadonlySpan of char and count number of characters consumed
+    /// </summary>
+    /// <param name="s"></param>
+    /// <param name="characters_consumed"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     public static unsafe float ParseFloat(ReadOnlySpan<char> s, out long characters_consumed, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
     {
       fixed (char* pStart = s)
@@ -90,26 +107,46 @@ namespace csFastFloat
         return ParseNumber(pStart, pStart + (uint)s.Length, out characters_consumed, expectedFormat, decimal_separator);
       }
     }
+
+
+    /// <summary>
+    /// Parse float between two char pointers
+    /// </summary>
+    /// <param name="first"></param>
+    /// <param name="last"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     unsafe static public float ParseFloat(char* first, char* last, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
       => ParseNumber(first, last, out long _, expectedFormat, decimal_separator);
 
-
+    /// <summary>
+    /// Parse flaot between two char pointers and count number of characters consumed
+    /// </summary>
+    /// <param name="first"></param>
+    /// <param name="last"></param>
+    /// <param name="characters_consumed"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     unsafe static internal float ParseNumber(char* first, char* last, out long characters_consumed, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
     {
+      var leading_spaces = 0;
       while ((first != last) && Utils.is_space((byte)(*first)))
       {
         first++;
+        leading_spaces++;
       }
       if (first == last)
       {
         ThrowArgumentException();
       }
-      ParsedNumberString pns = ParseNumberString(first, last, expectedFormat);
+      ParsedNumberString pns = ParsedNumberString.ParseNumberString(first, last, expectedFormat);
       if (!pns.valid)
       {
         return HandleInvalidInput(first, last, out characters_consumed);
       }
-      characters_consumed = pns.characters_consumed;
+      characters_consumed = pns.characters_consumed + leading_spaces;
 
       // Next is Clinger's fast path.
       if (FloatBinaryConstants.min_exponent_fast_path <= pns.exponent && pns.exponent <= FloatBinaryConstants.max_exponent_fast_path && pns.mantissa <= FloatBinaryConstants.max_mantissa_fast_path && !pns.too_many_digits)
@@ -132,6 +169,15 @@ namespace csFastFloat
     }
 
 
+    /// <summary>
+    /// Parse float between two byte pointers and count number of characters consumed (UTF-8)
+    /// </summary>
+    /// <param name="first"></param>
+    /// <param name="last"></param>
+    /// <param name="characters_consumed"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     unsafe static internal float ParseNumber(byte* first, byte* last, out long characters_consumed, chars_format expectedFormat = chars_format.is_general, byte decimal_separator = (byte)'.')
     {
       var leading_spaces = 0;
@@ -145,7 +191,7 @@ namespace csFastFloat
       {
         ThrowArgumentException();
       }
-      ParsedNumberString pns = ParseNumberString(first, last, expectedFormat);
+      ParsedNumberString pns = ParsedNumberString.ParseNumberString(first, last, expectedFormat);
       if (!pns.valid)
       {
         return HandleInvalidInput(first, last, out characters_consumed);
@@ -171,13 +217,28 @@ namespace csFastFloat
       if (am.power2 < 0) { am = ParseLongMantissa(first, last, (byte)decimal_separator); }
       return ToFloat(pns.negative, am);
     }
+    
+
+    /// <summary>
+    /// Parse float from ReadOnly span of bytes (UTF-8)
+    /// </summary>
+    /// <param name="s"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     public static unsafe float ParseFloat(ReadOnlySpan<byte> s, chars_format expectedFormat = chars_format.is_general, byte decimal_separator = (byte)'.')
-    {
-      fixed(byte* pStart = s)
-      {
-        return ParseNumber(pStart, pStart + s.Length, out long _, expectedFormat, decimal_separator);
-      }
-    }
+      => ParseFloat(s, out long _, expectedFormat, decimal_separator);
+
+
+
+    /// <summary>
+    /// Parse float from ReadOnly span of bytes (UTF-8) and count number of characters consumed
+    /// </summary>
+    /// <param name="s"></param>
+    /// <param name="characters_consumed"></param>
+    /// <param name="expectedFormat"></param>
+    /// <param name="decimal_separator"></param>
+    /// <returns></returns>
     public static unsafe float ParseFloat(ReadOnlySpan<byte> s, out long characters_consumed, chars_format expectedFormat = chars_format.is_general, byte decimal_separator = (byte)'.')
     {
       fixed(byte* pStart = s)
@@ -185,14 +246,14 @@ namespace csFastFloat
         return ParseNumber(pStart, pStart + s.Length, out characters_consumed, expectedFormat, decimal_separator);
       }
     }
+    
     /// <summary>
-    ///
+    /// Compute adjusted mantissa
     /// </summary>
-    /// <param name="q"></param>
-    /// <param name="w"></param>
+    /// <param name="q">exponent</param>
+    /// <param name="w">significand</param>
     ///
     /// <returns></returns>
-
     internal static AdjustedMantissa ComputeFloat(long q, ulong w)
     {
       var answer = new AdjustedMantissa();
@@ -309,7 +370,11 @@ namespace csFastFloat
 
 
 
-
+    /// <summary>
+    /// Compute adjusted mantissa from decimal info
+    /// </summary>
+    /// <param name="d"></param>
+    /// <returns></returns>
     internal static AdjustedMantissa ComputeFloat(DecimalInfo d)
     {
       AdjustedMantissa answer = new AdjustedMantissa();
@@ -563,320 +628,7 @@ namespace csFastFloat
 
 
 
-    unsafe static internal ParsedNumberString ParseNumberString(char* p, char* pend, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
-    {
-      ParsedNumberString answer = new ParsedNumberString();
-
-      answer.valid = false;
-      answer.too_many_digits = false;
-      char* pstart = p;
-      answer.negative = (*p == '-');
-      if ((*p == '-') || (*p == '+'))
-      {
-        ++p;
-        if (p == pend)
-        {
-          return answer;
-        }
-        if (!Utils.is_integer(*p, out uint _) && (*p != decimal_separator)) // culture info ?
-        { // a  sign must be followed by an integer or the dot
-          return answer;
-        }
-      }
-      char* start_digits = p;
-
-      ulong i = 0; // an unsigned int avoids signed overflows (which are bad)
-
-      while ((p != pend) && Utils.is_integer(*p, out uint cMinus0))
-      {
-        // a multiplication by 10 is cheaper than an arbitrary integer
-        // multiplication
-        i = 10 * i + (ulong)cMinus0; // might overflow, we will handle the overflow later
-        ++p;
-      }
-      char* end_of_integer_part = p;
-      long digit_count = (long)(end_of_integer_part - start_digits);
-      long exponent = 0;
-      if ((p != pend) && (*p == decimal_separator))
-      {
-        ++p;
-        while ((p != pend) && Utils.is_integer(*p, out uint cMinus0))
-        {
-          byte digit = (byte)cMinus0;
-          ++p;
-          i = i * 10 + digit; // in rare cases, this will overflow, but that's ok
-        }
-        exponent = end_of_integer_part + 1 - p;
-        digit_count -= exponent;
-      }
-      // we must have encountered at least one integer!
-      if (digit_count == 0)
-      {
-        return answer;
-      }
-      long exp_number = 0;            // explicit exponential part
-      if (expectedFormat.HasFlag(chars_format.is_scientific) && (p != pend) && (('e' == *p) || ('E' == *p)))
-      {
-        char* location_of_e = p;
-        ++p;
-        bool neg_exp = false;
-        if ((p != pend) && ('-' == *p))
-        {
-          neg_exp = true;
-          ++p;
-        }
-        else if ((p != pend) && ('+' == *p))
-        {
-          ++p;
-        }
-        if ((p == pend) || !Utils.is_integer(*p, out uint _))
-        {
-          if (expectedFormat != chars_format.is_fixed)
-          {
-            // We are in error.
-            return answer;
-          }
-          // Otherwise, we will be ignoring the 'e'.
-          p = location_of_e;
-        }
-        else
-        {
-          while ((p != pend) && Utils.is_integer(*p, out uint cMinus0))
-          {
-            byte digit = (byte)cMinus0;
-            if (exp_number < 0x10000)
-            {
-              exp_number = 10 * exp_number + digit;
-            }
-            ++p;
-          }
-          if (neg_exp) { exp_number = -exp_number; }
-          exponent += exp_number;
-        }
-      }
-      else
-      {
-        // If it scientific and not fixed, we have to bail out.
-        if ((expectedFormat.HasFlag(chars_format.is_scientific)) && !(expectedFormat.HasFlag(chars_format.is_fixed))) { return answer; }
-      }
-      answer.valid = true;
-      answer.characters_consumed = p - pstart;
-
-      // If we frequently had to deal with long strings of digits,
-      // we could extend our code by using a 128-bit integer instead
-      // of a 64-bit integer. However, this is uncommon.
-      //
-      // We can deal with up to 19 digits.
-      if (digit_count > 19)
-      { // this is uncommon
-        // It is possible that the integer had an overflow.
-        // We have to handle the case where we have 0.0000somenumber.
-        // We need to be mindful of the case where we only have zeroes...
-        // E.g., 0.000000000...000.
-        char* start = start_digits;
-        while ((start != pend) && (*start == '0' || *start == decimal_separator))
-        {
-          if (*start == '0') { digit_count--; }
-          start++;
-        }
-        if (digit_count > 19)
-        {
-          answer.too_many_digits = true;
-          // Let us start again, this time, avoiding overflows.
-          i = 0;
-          p = start_digits;
-          const ulong minimal_nineteen_digit_integer = 1000000000000000000;
-          while ((i < minimal_nineteen_digit_integer) && (p != pend) && Utils.is_integer(*p, out uint cMinus0))
-          {
-            i = i * 10 + (ulong)cMinus0;
-            ++p;
-          }
-          if (i >= minimal_nineteen_digit_integer)
-          { // We have a big integers
-            exponent = end_of_integer_part - p + exp_number;
-          }
-          else
-          { // We have a value with a fractional component.
-            p++; // skip the '.'
-            char* first_after_period = p;
-            while ((i < minimal_nineteen_digit_integer) && (p != pend) && Utils.is_integer(*p, out uint cMinus0))
-            {
-              i = i * 10 + (ulong)cMinus0;
-              ++p;
-            }
-            exponent = first_after_period - p + exp_number;
-          }
-          // We have now corrected both exponent and i, to a truncated value
-        }
-      }
-      answer.exponent = exponent;
-      answer.mantissa = i;
-      return answer;
-    }
-
-
-    unsafe static internal ParsedNumberString ParseNumberString(byte* p, byte* pend, chars_format expectedFormat = chars_format.is_general, char decimal_separator = '.')
-    {
-      ParsedNumberString answer = new ParsedNumberString();
-
-      answer.valid = false;
-      answer.too_many_digits = false;
-      byte* pstart = p;
-      answer.negative = (*p == '-');
-      if ((*p == '-') || (*p == '+'))
-      {
-        ++p;
-        if (p == pend)
-        {
-          return answer;
-        }
-        if (!Utils.is_integer(*p, out uint digit) && (*p != decimal_separator)) // culture info ?
-        { // a  sign must be followed by an integer or the dot
-          return answer;
-        }
-      }
-      byte* start_digits = p;
-
-      ulong i = 0; // an unsigned int avoids signed overflows (which are bad)
-
-      while ((p != pend) && Utils.is_integer(*p, out uint digit))
-      {
-        // a multiplication by 10 is cheaper than an arbitrary integer
-        // multiplication
-        i = 10 * i + digit; // might overflow, we will handle the overflow later
-        ++p;
-      }
-      byte* end_of_integer_part = p;
-      long digit_count = (long)(end_of_integer_part - start_digits);
-      long exponent = 0;
-      if ((p != pend) && (*p == decimal_separator))
-      {
-        ++p;
-        if ((p + 8 <= pend) && Utils.is_made_of_eight_digits_fast(p))
-        {
-          i = i * 100000000 + Utils.parse_eight_digits_unrolled(p);
-          p += 8;
-          if ((p + 8 <= pend) && Utils.is_made_of_eight_digits_fast(p)) {
-            i = i * 100000000 + Utils.parse_eight_digits_unrolled(p);
-            p += 8;
-          }
-        }
-        while ((p != pend) && Utils.is_integer(*p, out uint digit))
-        {
-          ++p;
-          i = i * 10 + digit; // in rare cases, this will overflow, but that's ok
-        }
-        exponent = end_of_integer_part + 1 - p;
-        digit_count -= exponent;
-      }
-      // we must have encountered at least one integer!
-      if (digit_count == 0)
-      {
-        return answer;
-      }
-      long exp_number = 0;            // explicit exponential part
-      if (expectedFormat.HasFlag(chars_format.is_scientific) && (p != pend) && (('e' == *p) || ('E' == *p)))
-      {
-        byte* location_of_e = p;
-        ++p;
-        bool neg_exp = false;
-        if ((p != pend) && ('-' == *p))
-        {
-          neg_exp = true;
-          ++p;
-        }
-        else if ((p != pend) && ('+' == *p))
-        {
-          ++p;
-        }
-        if ((p == pend) || !Utils.is_integer(*p, out uint digit))
-        {
-          if (expectedFormat != chars_format.is_fixed)
-          {
-            // We are in error.
-            return answer;
-          }
-          // Otherwise, we will be ignoring the 'e'.
-          p = location_of_e;
-        }
-        else
-        {
-          while ((p != pend) && Utils.is_integer(*p, out uint cdigit))
-          {
-            if (exp_number < 0x10000)
-            {
-              exp_number = 10 * exp_number + cdigit;
-            }
-            ++p;
-          }
-          if (neg_exp) { exp_number = -exp_number; }
-          exponent += exp_number;
-        }
-      }
-      else
-      {
-        // If it scientific and not fixed, we have to bail out.
-        if ((expectedFormat.HasFlag(chars_format.is_scientific)) && !(expectedFormat.HasFlag(chars_format.is_fixed))) { return answer; }
-      }
-      answer.valid = true;
-      answer.characters_consumed = p - pstart;
-
-      // If we frequently had to deal with long strings of digits,
-      // we could extend our code by using a 128-bit integer instead
-      // of a 64-bit integer. However, this is uncommon.
-      //
-      // We can deal with up to 19 digits.
-      if (digit_count > 19)
-      { // this is uncommon
-        // It is possible that the integer had an overflow.
-        // We have to handle the case where we have 0.0000somenumber.
-        // We need to be mindful of the case where we only have zeroes...
-        // E.g., 0.000000000...000.
-        byte* start = start_digits;
-        while ((start != pend) && (*start == '0' || *start == decimal_separator))
-        {
-          if (*start == '0') { digit_count--; }
-          start++;
-        }
-        if (digit_count > 19)
-        {
-          answer.too_many_digits = true;
-          // Let us start again, this time, avoiding overflows.
-          i = 0;
-          p = start_digits;
-          const ulong minimal_nineteen_digit_integer = 1000000000000000000;
-          while ((i < minimal_nineteen_digit_integer) && (p != pend) && Utils.is_integer(*p, out uint digit))
-          {
-            i = i * 10 + digit;
-            ++p;
-          }
-          if (i >= minimal_nineteen_digit_integer)
-          { // We have a big integers
-            exponent = end_of_integer_part - p + exp_number;
-          }
-          else
-          { // We have a value with a fractional component.
-            p++; // skip the '.'
-            byte* first_after_period = p;
-            while ((i < minimal_nineteen_digit_integer) && (p != pend) && Utils.is_integer(*p, out uint digit))
-            {
-              i = i * 10 + digit;
-              ++p;
-            }
-            exponent = first_after_period - p + exp_number;
-          }
-          // We have now corrected both exponent and i, to a truncated value
-        }
-      }
-      answer.exponent = exponent;
-      answer.mantissa = i;
-      return answer;
-    }
-
-    // This should always succeed since it follows a call to parse_number_string
-    // This function could be optimized. In particular, we could stop after 19 digits
-    // and try to bail out. Furthermore, we should be able to recover the computed
-    // exponent from the pass in parse_number_string.
+    
   }
